+++ RockLib.HealthChecks/CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## 4.0.0-alpha.1 - 2025-02-04
+
+#### Changed
+- Removed .NET 6 as a target framework
+
+#### Removed
+- RockLib.Immutable
+
 ## 3.0.2 - 2024-10-24
 
 #### Changed